ExcelDataExporter.Loader/ExcelSheetDataReader.cs
using ExcelDataReader;

namespace Physalia.ExcelDataExporter.Loader
{
    internal class ExcelSheetDataReader
    {
        public SheetRawData ReadSheet(IExcelDataReader reader)
        {
            int rowCount = reader.RowCount;
            int columnCount = reader.FieldCount;
            var sheetRawData = new SheetRawData(rowCount, columnCount);
            sheetRawData.SetName(reader.Name);

            var rowIndex = 0;
            while (reader.Read())
            {
                for (var columnIndex = 0; columnIndex < columnCount; columnIndex++)
                {
                    string text = reader.GetValue(columnIndex)?.ToString();
                    sheetRawData.Set(rowIndex, columnIndex, text);
                }

                rowIndex++;
            }

            sheetRawData.ResizeBounds();

            // Set metadata
            string metadataText = sheetRawData.Get(0, 0);
            sheetRawData.SetMetadata(metadataText);

            // Remove metadata row
            sheetRawData.RemoveRow(0);

            return sheetRawData;
        }
    }
}
